Fix accounts change event handling
		we've updated the ui watching the connected account change
		for pegin and pegout forms. Also the rlogin supported networks
		are enabled again.,
