+++ src/common/store/session/actions.ts
@@ -1,8 +1,4 @@
 import { ActionTree } from 'vuex';
-import RLogin from '@rsksmart/rlogin';
-import WalletConnectProvider from '@walletconnect/web3-provider';
-import { trezorProviderOptions } from '@rsksmart/rlogin-trezor-provider';
-import { ledgerProviderOptions } from '@rsksmart/rlogin-ledger-provider';
 import axios, { AxiosResponse } from 'axios';
 import * as constants from '@/common/store/constants';
 import {
@@ -10,74 +6,20 @@ import {
   AppLocale,
   FeatureNames,
 } from '@/common/types';
-import { EnvironmentAccessorService } from '@/common/services/enviroment-accessor.service';
 import { ApiService } from '@/common/services';
 import {
   getBtcAddressFromSignedMessage,
   getCookie,
+  getRloginInstance,
   setCookie,
 } from '@/common/utils';
 import { ethers, providers } from 'ethers';
 import { markRaw } from 'vue';
 import { toUtf8Bytes } from 'ethers/lib/utils';
 
-const rpcUrls = {};
-const supportedChains = Object.keys(rpcUrls).map(Number);
-const customLedgerProviderOptions = ledgerProviderOptions;
-customLedgerProviderOptions.connector = async (ProviderPackage, options) => {
-  const ledgerOptions = options;
-  ledgerOptions.messageHashed = true;
-  const provider = new ProviderPackage(ledgerOptions);
-  await provider.connect();
-  return provider;
-};
-const rLoginSetup = new RLogin({
-  cacheProvider: true,
-  defaultTheme: 'dark',
-  providerOptions: {
-    walletconnect: {
-      package: WalletConnectProvider,
-      options: {
-        rpc: rpcUrls,
-      },
-    },
-    'custom-ledger': customLedgerProviderOptions,
-    'custom-trezor': {
-      ...trezorProviderOptions,
-      options: {
-        dPath: "m/44'/37310'/0'/0/0",
-        manifestEmail: EnvironmentAccessorService
-          .getEnvironmentVariables().vueAppManifestEmail,
-        manifestAppUrl: EnvironmentAccessorService
-          .getEnvironmentVariables().vueAppManifestAppUrl,
-      },
-    },
-  },
-  rpcUrls,
-  supportedChains,
-});
-
 export const actions: ActionTree<SessionState, RootState> = {
   [constants.SESSION_CONNECT_WEB3]: ({ commit, state, dispatch }): Promise<void> => {
-    const network = EnvironmentAccessorService.getEnvironmentVariables().vueAppCoin;
-    if (network === constants.BTC_NETWORK_MAINNET) {
-      Object
-        .defineProperty(rpcUrls, constants.SUPPORTED_NETWORKS.RSK_MAINNET.chainId, {
-          value: constants.SUPPORTED_NETWORKS.RSK_MAINNET.rpcUrl,
-          writable: false,
-          configurable: true,
-          enumerable: true,
-        });
-    } else {
-      Object
-        .defineProperty(rpcUrls, constants.SUPPORTED_NETWORKS.RSK_TESTNET.chainId, {
-          value: constants.SUPPORTED_NETWORKS.RSK_TESTNET.rpcUrl,
-          writable: false,
-          configurable: true,
-          enumerable: true,
-        });
-    }
-    const rLogin = state.rLoginInstance === undefined ? rLoginSetup : state.rLoginInstance;
+    const rLogin = state.rLoginInstance === undefined ? getRloginInstance() : state.rLoginInstance;
     return new Promise<void>((resolve, reject) => {
       rLogin.connect()
         .then((rLoginResponse) => {
@@ -192,7 +134,7 @@ export const actions: ActionTree<SessionState, RootState> = {
     }
   },
   [constants.SESSION_CONNECT_WEB3_FROM_CACHE]: async ({ commit, dispatch }) => {
-    const rLogin = rLoginSetup;
+    const rLogin = getRloginInstance();
     rLogin.connectTo(rLogin.cachedProvider)
       .then((rLoginResponse) => {
         const provider = new providers.Web3Provider(rLoginResponse.provider);
