+++ src/common/utils/index.ts
@@ -6,4 +6,5 @@ export * from './btcAddressUtils';
 export * from './utils';
 export * from './clarity';
 export * from './ethers';
+export * from './rlogin';
 export { default as ServiceError } from './ServiceError';
